Refine meaning of null value for too large data sets
If a GET request wants to retrieve a too large data set (e.g. the
device's entire data), the device should not just answer with an
error that the response is too large, but instead try to provide as
much useful information in the response as possible.

Defining null as an indicator that the data exists, but cannot be
provided in this request helps to reduce the number of necessary
requests to discover an unknown device.

As a consequence, null should not be used as a return value for
an executable item without parameters. An empty array is more
suitable and consistent, so this was adjusted in the data structure.