CAN bus: Add automatic node address assignment
The node addresses (used in the source and target bytes) had to be
hard-coded so far. This commit introduces a an automatic node address
assignment process, which is similar to SAE J1939 address claiming
with some improvements to reduce collisions.

For efficient implementation in the firmware it is required to have
a dedicated CAN frame type for network management whose IDs don't
overlap with existing req/resp and pub/sub frames. For this reason,
another variant defined by the DP and EDP bits from SAE J1939 had
to be used.

The previous approach was to select bits which are currently not used
by SAE J1939 or NMEA2000. However, this would only leave 2 message
types for ThingSet.

As ThingSet uses 500 kbit/s as its fixed bitrate, which is incompatible
with above mentioned standards anyway, it was decided to drop the idea
of allowing to use the same bus for different standards and use the
SAE J1939 CAN ID space for ThingSet request/response messages.

It may actually still be possible to share a bus between ThingSet and
SAE J1939 after the node addresses have been assigned, but it's not
a goal of the ThingSet protocol (anymore) and the protocol will not
be constrained in order to support this.