Make physical quantities part of data object name
The previous concept of adding them on the application-side based
on the unit turned out to be not viable. For example time differences
and durations as well as absolute times are measured in seconds.
Letting the app add the pyshical quantity would make "Charging Delay
Time" out of "wChargingDelay_s".

Or a metre can be used for a length or a distance, so it is ambiguous
if "_m" should be translated to Distance or Length.

Also some units are not always used for the same physical quantity,
e.g. Pa can be used for a pressure as well as mechanical stress.

Adding the physical quantities to the data object name removes any
unambigousity and makes parsing easier (with the drawback of slightly
increasing the length of some object names).