Text mode: Use fetch request with null for discovery instead of ?/
Discovery in the text mode now uses a fetch request with JSON `null` as
the argument instead of a trailing `/`. This is more consistent with
the binary mode and allows better mapping between the two.
In addition to that it avoids confusion for `?/` requests, where it is
not obvious if it's a discovery request for the root path of a node or
an absolute path.