BREAKING: Text mode response: Remove status-msg and add node ID
In order to allow gateways to talk to multiple downstream nodes
simultaneously (before pending responses for other nodes arrived)
the responses for requests with absolute paths need to carry the
node ID.

This requires a breaking change to the response layout.

As the status message was unnecessary overhead, they are now aligned
with CoAP diagnostic payload: A detailed status message can be provided
as payload for otherwise empty error messages. Successful messages only
use the status code and no redundant status message.