Bluetooth: Change UUID and packetization spec
Switching to line-feed as the end character, which allows to use the
same packetization mechanism for bluetooth and serial in the future.

As the carriage-return character is ignored by the serial, it must
be skipped/escaped also in the packetization spec.

Escape characters are now chosen such that the propability of
needing them is reduced. Previously used `C0` is also used as a
response code in the protocol, which should be avoided.

In order to avoid incompatibilities with previous implementations,
a new UUID is chosen.