CAN: Fix packetization for multi-frame reports
With the previous approach it was not possible to the beginning of a message
when connected to a bus that is already in operation (e.g. after a reset or
during hot-plug).

This is not acceptable, so we need to add a first and last frame indication.

The message number is reduced from 3 to 2 bits, which is still considered
sufficient.

The order of the new 2-bit multi-frame type is such that a CAN controller
prioritizing frames in the queue by ID would always still send out frames
in the correct order. First frame has the highest priority, followed by
consecutive frame, last frame and single frame type.